make release.sh safer
if $DEST is empty, "rm -rf $DEST/*" expands to "rm -rf /*". bash
scripts can avoid this by using "${DEST:?}" which throws a "parameter
null or not set" error when $DEST is empty.

+++ release.sh
@@ -11,7 +11,7 @@ if [[ "$V" == "" ]] ; then
 else
     [[ -e "$ZIP" ]] && rm "$ZIP"
     [[ -f "$DEST" ]] || mkdir "$DEST"
-    rm -rf "$DEST/*"
+    rm -rf "${DEST:?}/*" # avoid rm /* if $DEST is empty. thanks shellcheck!
     for f in $FILES ; do cp -a "$f" "$DEST" ; done
     zip -r "$ZIP" "$DEST"
 fi
